refactor: Remove values from TableReference
