+++ src/query/sql/src/planner/binder/mod.rs
@@ -47,6 +47,7 @@ mod stage;
 mod table;
 mod table_args;
 mod update;
+mod values;
 mod window;
 
 pub use aggregate::AggregateInfo;
